Critical changes to files
Changed how the data is read when averaging SAXS kinetic data because the subtracted curves treated with SAXSutilities have an extra space at the end that causes issues with Pandas. A bit of a stupid workaround.

The find time between runs has an error in its calculation and, to be honest, if you have the time constants used with ccdmcal/ccdmvdc it's better to calculate them using a spreadsheet than using this.